pyproject.toml
[project]
name = "coverage-gap"
version = "0.1.0"
description = "Specialist Coverage Gap Index for Mississippi Critical Access Hospitals"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pandas>=2.2",
    "geopandas>=1.0",
    "shapely>=2.0",
    "requests>=2.32",
    "tenacity>=8.2",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pyarrow>=16.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
coverage-gap = "coverage_gap.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coverage_gap"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
